api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
azure-cosmos>=4.0.0
azure-identity
pytest
pytest-asyncio
aiohttp